mysite/requirements.txt
django
djangorestframework
environs
psycopg2
